requirements.txt
pandas==1.0.3
xlrd
pdfplumber==0.5.23
azure-functions
azure-storage-blob==12.3.2
pandas==1.0.3
xlrd
pdfplumber==0.5.23
numpy==1.18.4